Falta consertar ainda o tipo que a imagem é retornada do banco de dados

+++ app/src/main/java/com/example/alexandriavirtual20/adapter/ListaEventoAdapter.kt
@@ -34,7 +34,7 @@ class ListaEventoAdapter (
     override fun onBindViewHolder(holder: ListaEventoAdapter.ViewHolder, position: Int) {
         val evento = eventos[position]
 
-        holder.imagem.setImageResource(evento.imagem)
+        holder.imagem.setImageResource((evento.imagem).toInt())
         holder.nome.text = evento.nome
         holder.data.text = evento.data
         holder.descricao.text = evento.descricao
